+++ src/index.ts
@@ -1,1 +1,3 @@
+// sets up string formats
+import "./formats";
 export * from "./validation";