+++ changelog.md
@@ -1,3 +1,11 @@
+# 1.0.3
+
+- cache all schemas after first validation
+
+# 1.0.2
+
+- adds missing comment to validateData function
+
 # 1.0.1
 
 - fix wrong initial publish on npm, adapt paths in package.json